Only send top level elements changes

+++ editor/src/components/editor/store/collaborative-editing.ts
@@ -1,4 +1,4 @@
-import type { CollabTextFile } from './editor-state'
+import type { CollabTextFile, CollabTextFileTopLevelElements } from './editor-state'
 import { type CollaborativeEditingSupport } from './editor-state'
 import {
   writeProjectFileChange,
@@ -16,16 +16,23 @@ import {
   zipContentsTree,
 } from '../../../components/assets'
 import type { EditorDispatch } from '../action-types'
-import { applyCollabFileUpdate, updateFile } from '../actions/action-creators'
+import {
+  applyCollabFileUpdate,
+  updateFile,
+  updateTopLevelElements,
+} from '../actions/action-creators'
 import { forceNotNull } from '../../../core/shared/optional-utils'
 import { assertNever, isBrowserEnvironment } from '../../../core/shared/utils'
 import type { ParseSuccess } from '../../../core/shared/project-file-types'
 import { isTextFile } from '../../../core/shared/project-file-types'
 import {
   ParsedTextFileKeepDeepEquality,
+  TextFileKeepDeepEquality,
   TopLevelElementKeepDeepEquality,
 } from './store-deep-equality-instances'
 import * as Y from 'yjs'
+import type { TopLevelElement } from '../../../core/shared/element-template'
+import { v4 as UUID } from 'uuid'
 
 export function collateCollaborativeProjectChanges(
   oldContents: ProjectContentTreeRoot,
@@ -118,6 +125,7 @@ export function collateCollaborativeProjectChanges(
 function applyFileChangeToMap(
   change: ProjectFileChange,
   projectContentsMap: CollaborativeEditingSupport['projectContents'],
+  mergeDoc: Y.Doc,
 ): void {
   switch (change.type) {
     case 'DELETE_PATH':
@@ -138,13 +146,24 @@ function applyFileChangeToMap(
         change.projectFile.type === 'TEXT_FILE' &&
         change.projectFile.fileContents.parsed.type === 'PARSE_SUCCESS'
       ) {
-        const collabFile: CollabTextFile = new Y.Map()
-        collabFile.set('type', 'TEXT_FILE')
-        collabFile.set(
-          'topLevelElements',
-          Y.Array.from(change.projectFile.fileContents.parsed.topLevelElements),
-        )
-        projectContentsMap.set(change.fullPath, collabFile)
+        const parsedPart = change.projectFile.fileContents.parsed
+        let collabFile: CollabTextFile
+        if (projectContentsMap.has(change.fullPath)) {
+          collabFile = projectContentsMap.get(change.fullPath)!
+          // TODO Handle not text type
+        } else {
+          collabFile = new Y.Map()
+          collabFile.set('type', 'TEXT_FILE')
+          const topLevelElementsArray = new Y.Array<TopLevelElement>()
+          collabFile.set('topLevelElements', topLevelElementsArray)
+        }
+
+        mergeDoc.transact(() => {
+          if (!projectContentsMap.has(change.fullPath)) {
+            projectContentsMap.set(change.fullPath, collabFile)
+          }
+          synchroniseParseSuccessToCollabFile(parsedPart, collabFile)
+        })
       }
       break
     case 'ENSURE_DIRECTORY_EXISTS':
@@ -155,49 +174,93 @@ function applyFileChangeToMap(
   }
 }
 
+// FIXME This is still being called on load
 export function updateCollaborativeProjectContents(
   collaborativeEditingSupport: CollaborativeEditingSupport,
   projectChanges: ProjectChanges,
 ): void {
   const projectContentsMap = collaborativeEditingSupport.projectContents
   for (const change of projectChanges.fileChanges.collabProjectChanges) {
-    applyFileChangeToMap(change, projectContentsMap)
+    applyFileChangeToMap(change, projectContentsMap, collaborativeEditingSupport.mergeDoc)
   }
 }
 
-export function populateCollaborativeProjectContents(
-  collaborativeEditingSupport: CollaborativeEditingSupport,
-  projectContents: ProjectContentTreeRoot,
-): void {
-  walkContentsTree(projectContents, (fullPath, file) => {
-    applyFileChangeToMap(
-      writeProjectFileChange(fullPath, file),
-      collaborativeEditingSupport.projectContents,
-    )
-  })
-}
+// export function populateCollaborativeProjectContents(
+//   collaborativeEditingSupport: CollaborativeEditingSupport,
+//   projectContents: ProjectContentTreeRoot,
+// ): void {
+//   walkContentsTree(projectContents, (fullPath, file) => {
+//     applyFileChangeToMap(
+//       writeProjectFileChange(fullPath, file),
+//       collaborativeEditingSupport.projectContents,
+//       collaborativeEditingSupport.mergeDoc,
+//     )
+//   })
+// }
 
 export function addHookForProjectChanges(
   collaborativeEditingSupport: CollaborativeEditingSupport,
   dispatch: EditorDispatch,
 ): void {
-  collaborativeEditingSupport.projectContents.observe((changeEvent) => {
-    changeEvent.changes.keys.forEach((change, key) => {
-      switch (change.action) {
-        case 'add':
-        case 'update':
-          const file = forceNotNull(
-            'Should not be null.',
-            collaborativeEditingSupport.projectContents.get(key),
-          )
-          dispatch([applyCollabFileUpdate(key, file)])
-          break
-        case 'delete':
-          console.error('Deletes to be handled later.')
-          break
-        default:
-          console.error(`Not sure what this is: ${change}`)
+  // collaborativeEditingSupport.projectContents.observe((changeEvent) => {
+  //   console.log(`observe`, changeEvent.changes)
+  //   changeEvent.changes.keys.forEach((change, key) => {
+  //     switch (change.action) {
+  //       case 'add':
+  //       case 'update':
+  //         const file = forceNotNull(
+  //           'Should not be null.',
+  //           collaborativeEditingSupport.projectContents.get(key),
+  //         )
+  //         dispatch([applyCollabFileUpdate(key, file)])
+  //         break
+  //       case 'delete':
+  //         console.error('Deletes to be handled later.')
+  //         break
+  //       default:
+  //         console.error(`Not sure what this is: ${change}`)
+  //     }
+  //   })
+  // })
+
+  collaborativeEditingSupport.projectContents.observeDeep((changeEvents) => {
+    // TODO Check that this is the array change before doing anything
+    changeEvents.forEach((changeEvent) => {
+      console.log(`observeDeep`, changeEvent.delta)
+      const filePath = changeEvent.path[0] as string
+      // FIXME Can be null actually
+      const file = forceNotNull(
+        'Should not be null.',
+        collaborativeEditingSupport.projectContents.get(filePath),
+      )
+      const oldTopLevelElements = file.get('topLevelElements') as CollabTextFileTopLevelElements
+      let newTopLevelElements: Array<TopLevelElement> = []
+      let writeIndex = 0
+      let readIndex = 0
+      changeEvent.delta.forEach((delta) => {
+        if (delta.retain != undefined) {
+          const elementsToPush = oldTopLevelElements.slice(readIndex, readIndex + delta.retain)
+          newTopLevelElements.push(...elementsToPush)
+          readIndex += delta.retain
+          writeIndex += delta.retain
+        }
+        if (delta.insert != null && Array.isArray(delta.insert)) {
+          newTopLevelElements.push(...delta.insert)
+          writeIndex += delta.insert.length
+        }
+        if (delta.delete != undefined) {
+          readIndex += delta.delete
+        }
+      })
+
+      if (readIndex < oldTopLevelElements.length) {
+        // There is an implicit retain for the remainder of the items
+        const elementsToPush = oldTopLevelElements.slice(readIndex)
+        newTopLevelElements.push(...elementsToPush)
       }
+
+      console.log(`newTopLevelElements`, newTopLevelElements)
+      dispatch([updateTopLevelElements(filePath, newTopLevelElements)])
     })
   })
 }
@@ -220,21 +283,21 @@ export function calculateArrayChanges<T>(
   equals: (from: T, into: T) => boolean = (fromToCheck, intoToCheck) => fromToCheck === intoToCheck,
 ): ArrayChanges {
   let updatesAt: Array<number> = []
-  for (let index: number = 0; index < Math.min(from.length, into.length); index++) {
+  for (let index: number = 0; index < Math.max(from.length, into.length); index++) {
     if (!equals(from[index], into.get(index))) {
       updatesAt.push(index)
     }
   }
 
   let deleteFrom: number | null = null
-  if (into.length < from.length) {
+  if (into.length > from.length) {
     deleteFrom = into.length
   }
 
   return arrayChanges(updatesAt, deleteFrom)
 }
 
-export function synchroniseParseSuccessToCollabFile(
+function synchroniseParseSuccessToCollabFile(
   success: ParseSuccess,
   collabFile: CollabTextFile,
 ): void {
@@ -242,9 +305,16 @@ export function synchroniseParseSuccessToCollabFile(
   if (collabFileTopLevelElements === 'TEXT_FILE' || collabFileTopLevelElements === undefined) {
     throw new Error('Invalid value for topLevelElements.')
   } else {
-    const changes = calculateArrayChanges(success.topLevelElements, collabFileTopLevelElements)
+    const changes = calculateArrayChanges(
+      success.topLevelElements,
+      collabFileTopLevelElements,
+      // Fix this shitshow up
+      (l, r) => l != null && r != null && TopLevelElementKeepDeepEquality(l, r).areEqual,
+    )
     for (const updateAtIndex of changes.updatesAt) {
-      collabFileTopLevelElements.delete(updateAtIndex, 1)
+      if (collabFileTopLevelElements.length > updateAtIndex) {
+        collabFileTopLevelElements.delete(updateAtIndex, 1)
+      }
       collabFileTopLevelElements.insert(updateAtIndex, [success.topLevelElements[updateAtIndex]])
     }
     if (changes.deleteFrom != null) {
